Extract stale review current-head evidence builders
